+++ Clyde.cpp
@@ -55,12 +55,12 @@ void Clyde::Update()
 			break;
 		}
 	case CHASE:					
-		mStart = new PathNode(this->mPos.x, this->mPos.z);
-		mGoal = new PathNode(round(MazeLoader::GetPacManData().at(0).pos.x), round(MazeLoader::GetPacManData().at(0).pos.z));
+		mStart = new PathNode((int)this->mPos.x, (int)this->mPos.z);
+		mGoal = new PathNode((int)round(MazeLoader::GetPacManData().at(0).pos.x), (int)round(MazeLoader::GetPacManData().at(0).pos.z));
 		waypoints = test.FindPath(mStart, mGoal);
 		if (waypoints.size() != 0)
 		{
-			this->setPos(XMVectorSet(waypoints.front()->xPos, 0.0f, waypoints.front()->zPos, 0.0f));
+			this->setPos(XMVectorSet((float)waypoints.front()->xPos, 0.0f, (float)waypoints.front()->zPos, 0.0f));
 		}		
 		break;
 	case FRIGHTENED:
