+++ Inky.cpp
@@ -41,19 +41,19 @@ void Inky::Update(float dt, PuckMan::Facing facingState, XMFLOAT3 blinkyPos)
 		if (facingState == PuckMan::Facing::F_FORWARD)
 		{
 			//Offset tile = 2 spaces in PuckMan's facing
-			offsetTile = new PathNode(round(MazeLoader::GetPacManData().at(0).pos.x), round(MazeLoader::GetPacManData().at(0).pos.z + 2));
+			offsetTile = new PathNode((int)round(MazeLoader::GetPacManData().at(0).pos.x), (int)round(MazeLoader::GetPacManData().at(0).pos.z + 2));
 			//Draw a vector from Blinky's current position to the offset tile's position
 			XMVECTOR targetTile = XMVectorSet(blinkyPos.x - offsetTile->xPos, 0.0f, blinkyPos.z - offsetTile->zPos, 0.0f);
 			//Double the vector length extending forward, this is Inky's target
 			targetTile = XMVectorScale(targetTile, 2.0f);
 
-			mStart = new PathNode(this->mPos.x, this->mPos.z);
-			mGoal = new PathNode(targetTile.m128_f32[0], targetTile.m128_f32[2]);
+			mStart = new PathNode((int)this->mPos.x, (int)this->mPos.z);
+			mGoal = new PathNode((int)targetTile.m128_f32[0], (int)targetTile.m128_f32[2]);
 			waypoints = test.FindPath(mStart, mGoal);
 
 			if (waypoints.size() != 0)
 			{
-				this->setPos(XMVectorSet(waypoints.front()->xPos, mPos.y, waypoints.front()->zPos, 0.0f));
+				this->setPos(XMVectorSet((float)waypoints.front()->xPos, mPos.y, (float)waypoints.front()->zPos, 0.0f));
 			}
 
 			break;
@@ -62,19 +62,19 @@ void Inky::Update(float dt, PuckMan::Facing facingState, XMFLOAT3 blinkyPos)
 		if (facingState == PuckMan::Facing::F_BACKWARD)
 		{
 			//Offset tile = 2 spaces in PuckMan's facing
-			offsetTile = new PathNode(round(MazeLoader::GetPacManData().at(0).pos.x), round(MazeLoader::GetPacManData().at(0).pos.z - 2));
+			offsetTile = new PathNode((int)round(MazeLoader::GetPacManData().at(0).pos.x), (int)round(MazeLoader::GetPacManData().at(0).pos.z - 2));
 			//Draw a vector from Blinky's current position to the offset tile's position
 			XMVECTOR targetTile = XMVectorSet(blinkyPos.x - offsetTile->xPos, 0.0f, blinkyPos.z - offsetTile->zPos, 0.0f);
 			//Double the vector length extending forward, this is Inky's target
 			targetTile = XMVectorScale(targetTile, 2.0f);
 
-			mStart = new PathNode(this->mPos.x, this->mPos.z);
-			mGoal = new PathNode(round(MazeLoader::GetPacManData().at(0).pos.x), round(MazeLoader::GetPacManData().at(0).pos.z - 4));
+			mStart = new PathNode((int)this->mPos.x, (int)this->mPos.z);
+			mGoal = new PathNode((int)round(MazeLoader::GetPacManData().at(0).pos.x), (int)round(MazeLoader::GetPacManData().at(0).pos.z - 4));
 			waypoints = test.FindPath(mStart, mGoal);
 
 			if (waypoints.size() != 0)
 			{
-				this->setPos(XMVectorSet(waypoints.front()->xPos, mPos.y, waypoints.front()->zPos, 0.0f));
+				this->setPos(XMVectorSet((float)waypoints.front()->xPos, mPos.y, (float)waypoints.front()->zPos, 0.0f));
 			}
 
 			break;
@@ -83,20 +83,20 @@ void Inky::Update(float dt, PuckMan::Facing facingState, XMFLOAT3 blinkyPos)
 		if (facingState == PuckMan::Facing::F_LEFT)
 		{
 			//Offset tile = 2 spaces in PuckMan's facing
-			offsetTile = new PathNode(round(MazeLoader::GetPacManData().at(0).pos.x - 2), round(MazeLoader::GetPacManData().at(0).pos.z));
+			offsetTile = new PathNode((int)round(MazeLoader::GetPacManData().at(0).pos.x - 2), (int)round(MazeLoader::GetPacManData().at(0).pos.z));
 			//Draw a vector from Blinky's current position to the offset tile's position
 			XMVECTOR targetTile = XMVectorSet(blinkyPos.x - offsetTile->xPos, 0.0f, blinkyPos.z - offsetTile->zPos, 0.0f);
 			//Double the vector length extending forward, this is Inky's target
 			targetTile = XMVectorScale(targetTile, 2.0f);
 
 
-			mStart = new PathNode(this->mPos.x, this->mPos.z);
-			mGoal = new PathNode(round(MazeLoader::GetPacManData().at(0).pos.x - 4), round(MazeLoader::GetPacManData().at(0).pos.z));
+			mStart = new PathNode((int)this->mPos.x, (int)this->mPos.z);
+			mGoal = new PathNode((int)round(MazeLoader::GetPacManData().at(0).pos.x - 4), (int)round(MazeLoader::GetPacManData().at(0).pos.z));
 			waypoints = test.FindPath(mStart, mGoal);
 
 			if (waypoints.size() != 0)
 			{
-				this->setPos(XMVectorSet(waypoints.front()->xPos, mPos.y, waypoints.front()->zPos, 0.0f));
+				this->setPos(XMVectorSet((float)waypoints.front()->xPos, mPos.y, (float)waypoints.front()->zPos, 0.0f));
 			}
 
 			break;
@@ -105,20 +105,20 @@ void Inky::Update(float dt, PuckMan::Facing facingState, XMFLOAT3 blinkyPos)
 		if (facingState == PuckMan::Facing::F_RIGHT)
 		{
 			//Offset tile = 2 spaces in PuckMan's facing
-			offsetTile = new PathNode(round(MazeLoader::GetPacManData().at(0).pos.x + 2), round(MazeLoader::GetPacManData().at(0).pos.z));
+			offsetTile = new PathNode((int)round(MazeLoader::GetPacManData().at(0).pos.x + 2), (int)round(MazeLoader::GetPacManData().at(0).pos.z));
 			//Draw a vector from Blinky's current position to the offset tile's position
 			XMVECTOR targetTile = XMVectorSet(blinkyPos.x - offsetTile->xPos, 0.0f, blinkyPos.z - offsetTile->zPos, 0.0f);
 			//Double the vector length extending forward, this is Inky's target
 			targetTile = XMVectorScale(targetTile, 2.0f);
 
 
-			mStart = new PathNode(this->mPos.x, this->mPos.z);
-			mGoal = new PathNode(round(MazeLoader::GetPacManData().at(0).pos.x + 4), round(MazeLoader::GetPacManData().at(0).pos.z));
+			mStart = new PathNode((int)this->mPos.x, (int)this->mPos.z);
+			mGoal = new PathNode((int)round(MazeLoader::GetPacManData().at(0).pos.x + 4), (int)round(MazeLoader::GetPacManData().at(0).pos.z));
 			waypoints = test.FindPath(mStart, mGoal);
 
 			if (waypoints.size() != 0)
 			{
-				this->setPos(XMVectorSet(waypoints.front()->xPos, mPos.y, waypoints.front()->zPos, 0.0f));
+				this->setPos(XMVectorSet((float)waypoints.front()->xPos, mPos.y, (float)waypoints.front()->zPos, 0.0f));
 			}
 
 			break;
